Added Sprache to support templates
We need to be able to specify some
template in form of simple text
files, to specify the layout of the
final document.

Using Sprache we can simply define
the syntax of this document and have
a parser that will create the real
C# composer class

In this first version we are able to simply
add static file and add work items from
a query.